Move phone location actions to its own module
