+++ src/passwordless/actions.js
@@ -3,39 +3,9 @@ import { read, getEntity, swap, updateEntity } from '../store/index';
 import { closeLock } from '../lock/actions';
 import webApi from '../lock/web_api';
 import * as c from '../cred/index';
-import * as cc from '../cred/country_codes';
 import * as cs from '../cred/storage';
 import * as l from '../lock/index';
 import * as m from './index';
-import {
-  closeLocationSelect,
-  openLocationSelect
-} from '../cred/phone-number/index';
-
-export function changePhoneLocation(id, location) {
-  swap(updateEntity, "lock", id, lock => {
-    lock = closeLocationSelect(lock);
-    lock = c.setPhoneLocation(lock, location);
-    return lock;
-  });
-}
-
-export function setDefaultLocation(id, str) {
-  const result = cc.findByIsoCode(str);
-  if (!result) {
-    throw new Error(`Unable to set the default location, can't find any country with the code "${str}".`);
-  }
-
-  swap(updateEntity, "lock", id, c.setPhoneLocation, result);
-}
-
-export function selectPhoneLocation(id, searchStr) {
-  swap(updateEntity, "lock", id, openLocationSelect, searchStr);
-}
-
-export function cancelSelectPhoneLocation(id) {
-  swap(updateEntity, "lock", id, closeLocationSelect);
-}
 
 export function requestPasswordlessEmail(id) {
   // TODO: abstract this submit thing.
